+++ executor/aggfuncs/func_avg.go
@@ -20,7 +20,6 @@ import (
 	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util/chunk"
 	"github.com/pingcap/tidb/util/hack"
-	"github.com/pingcap/tidb/util/set"
 )
 
 const (
@@ -193,7 +192,7 @@ func (e *avgPartial4Decimal) MergePartialResult(sctx sessionctx.Context, src, ds
 
 type partialResult4AvgDistinctDecimal struct {
 	partialResult4AvgDecimal
-	valSet set.StringSetWithMemoryUsage
+	valSet stringSetWithMemoryUsage
 }
 
 type avgOriginal4DistinctDecimal struct {
@@ -201,7 +200,7 @@ type avgOriginal4DistinctDecimal struct {
 }
 
 func (e *avgOriginal4DistinctDecimal) AllocPartialResult() (pr PartialResult, memDelta int64) {
-	valSet, memDelta := set.NewStringSetWithMemoryUsage()
+	valSet, memDelta := newStringSetWithMemoryUsage()
 	p := &partialResult4AvgDistinctDecimal{
 		valSet: valSet,
 	}
@@ -212,7 +211,7 @@ func (e *avgOriginal4DistinctDecimal) ResetPartialResult(pr PartialResult) {
 	p := (*partialResult4AvgDistinctDecimal)(pr)
 	p.sum = *types.NewDecFromInt(0)
 	p.count = int64(0)
-	p.valSet, _ = set.NewStringSetWithMemoryUsage()
+	p.valSet, _ = newStringSetWithMemoryUsage()
 }
 
 func (e *avgOriginal4DistinctDecimal) UpdatePartialResult(sctx sessionctx.Context, rowsInGroup []chunk.Row, pr PartialResult) (memDelta int64, err error) {
@@ -390,7 +389,7 @@ func (e *avgPartial4Float64) MergePartialResult(sctx sessionctx.Context, src, ds
 
 type partialResult4AvgDistinctFloat64 struct {
 	partialResult4AvgFloat64
-	valSet set.Float64SetWithMemoryUsage
+	valSet float64SetWithMemoryUsage
 }
 
 type avgOriginal4DistinctFloat64 struct {
@@ -398,7 +397,7 @@ type avgOriginal4DistinctFloat64 struct {
 }
 
 func (e *avgOriginal4DistinctFloat64) AllocPartialResult() (pr PartialResult, memDelta int64) {
-	valSet, memDelta := set.NewFloat64SetWithMemoryUsage()
+	valSet, memDelta := newFloat64SetWithMemoryUsage()
 	p := &partialResult4AvgDistinctFloat64{
 		valSet: valSet,
 	}
@@ -409,7 +408,7 @@ func (e *avgOriginal4DistinctFloat64) ResetPartialResult(pr PartialResult) {
 	p := (*partialResult4AvgDistinctFloat64)(pr)
 	p.sum = float64(0)
 	p.count = int64(0)
-	p.valSet, _ = set.NewFloat64SetWithMemoryUsage()
+	p.valSet, _ = newFloat64SetWithMemoryUsage()
 }
 
 func (e *avgOriginal4DistinctFloat64) UpdatePartialResult(sctx sessionctx.Context, rowsInGroup []chunk.Row, pr PartialResult) (memDelta int64, err error) {
