+++ executor/aggfuncs/func_count_distinct.go
@@ -28,7 +28,6 @@ import (
 	"github.com/pingcap/tidb/util/codec"
 	"github.com/pingcap/tidb/util/collate"
 	"github.com/pingcap/tidb/util/hack"
-	"github.com/pingcap/tidb/util/set"
 	"github.com/pingcap/tidb/util/stringutil"
 )
 
@@ -50,7 +49,7 @@ const (
 )
 
 type partialResult4CountDistinctInt struct {
-	valSet set.Int64SetWithMemoryUsage
+	valSet int64SetWithMemoryUsage
 }
 
 type countOriginalWithDistinct4Int struct {
@@ -58,7 +57,7 @@ type countOriginalWithDistinct4Int struct {
 }
 
 func (e *countOriginalWithDistinct4Int) AllocPartialResult() (pr PartialResult, memDelta int64) {
-	valSet, memDelta := set.NewInt64SetWithMemoryUsage()
+	valSet, memDelta := newInt64SetWithMemoryUsage()
 	return PartialResult(&partialResult4CountDistinctInt{
 		valSet: valSet,
 	}), DefPartialResult4CountDistinctIntSize + memDelta
@@ -66,7 +65,7 @@ func (e *countOriginalWithDistinct4Int) AllocPartialResult() (pr PartialResult, 
 
 func (e *countOriginalWithDistinct4Int) ResetPartialResult(pr PartialResult) {
 	p := (*partialResult4CountDistinctInt)(pr)
-	p.valSet, _ = set.NewInt64SetWithMemoryUsage()
+	p.valSet, _ = newInt64SetWithMemoryUsage()
 }
 
 func (e *countOriginalWithDistinct4Int) AppendFinalResult2Chunk(sctx sessionctx.Context, pr PartialResult, chk *chunk.Chunk) error {
@@ -97,7 +96,7 @@ func (e *countOriginalWithDistinct4Int) UpdatePartialResult(sctx sessionctx.Cont
 }
 
 type partialResult4CountDistinctReal struct {
-	valSet set.Float64SetWithMemoryUsage
+	valSet float64SetWithMemoryUsage
 }
 
 type countOriginalWithDistinct4Real struct {
@@ -105,7 +104,7 @@ type countOriginalWithDistinct4Real struct {
 }
 
 func (e *countOriginalWithDistinct4Real) AllocPartialResult() (pr PartialResult, memDelta int64) {
-	valSet, memDelta := set.NewFloat64SetWithMemoryUsage()
+	valSet, memDelta := newFloat64SetWithMemoryUsage()
 	return PartialResult(&partialResult4CountDistinctReal{
 		valSet: valSet,
 	}), DefPartialResult4CountDistinctRealSize + memDelta
@@ -113,7 +112,7 @@ func (e *countOriginalWithDistinct4Real) AllocPartialResult() (pr PartialResult,
 
 func (e *countOriginalWithDistinct4Real) ResetPartialResult(pr PartialResult) {
 	p := (*partialResult4CountDistinctReal)(pr)
-	p.valSet, _ = set.NewFloat64SetWithMemoryUsage()
+	p.valSet, _ = newFloat64SetWithMemoryUsage()
 }
 
 func (e *countOriginalWithDistinct4Real) AppendFinalResult2Chunk(sctx sessionctx.Context, pr PartialResult, chk *chunk.Chunk) error {
@@ -144,7 +143,7 @@ func (e *countOriginalWithDistinct4Real) UpdatePartialResult(sctx sessionctx.Con
 }
 
 type partialResult4CountDistinctDecimal struct {
-	valSet set.StringSetWithMemoryUsage
+	valSet stringSetWithMemoryUsage
 }
 
 type countOriginalWithDistinct4Decimal struct {
@@ -152,7 +151,7 @@ type countOriginalWithDistinct4Decimal struct {
 }
 
 func (e *countOriginalWithDistinct4Decimal) AllocPartialResult() (pr PartialResult, memDelta int64) {
-	valSet, memDelta := set.NewStringSetWithMemoryUsage()
+	valSet, memDelta := newStringSetWithMemoryUsage()
 	return PartialResult(&partialResult4CountDistinctDecimal{
 		valSet: valSet,
 	}), DefPartialResult4CountDistinctDecimalSize + memDelta
@@ -160,7 +159,7 @@ func (e *countOriginalWithDistinct4Decimal) AllocPartialResult() (pr PartialResu
 
 func (e *countOriginalWithDistinct4Decimal) ResetPartialResult(pr PartialResult) {
 	p := (*partialResult4CountDistinctDecimal)(pr)
-	p.valSet, _ = set.NewStringSetWithMemoryUsage()
+	p.valSet, _ = newStringSetWithMemoryUsage()
 }
 
 func (e *countOriginalWithDistinct4Decimal) AppendFinalResult2Chunk(sctx sessionctx.Context, pr PartialResult, chk *chunk.Chunk) error {
@@ -196,7 +195,7 @@ func (e *countOriginalWithDistinct4Decimal) UpdatePartialResult(sctx sessionctx.
 }
 
 type partialResult4CountDistinctDuration struct {
-	valSet set.Int64SetWithMemoryUsage
+	valSet int64SetWithMemoryUsage
 }
 
 type countOriginalWithDistinct4Duration struct {
@@ -204,7 +203,7 @@ type countOriginalWithDistinct4Duration struct {
 }
 
 func (e *countOriginalWithDistinct4Duration) AllocPartialResult() (pr PartialResult, memDelta int64) {
-	valSet, memDelta := set.NewInt64SetWithMemoryUsage()
+	valSet, memDelta := newInt64SetWithMemoryUsage()
 	return PartialResult(&partialResult4CountDistinctDuration{
 		valSet: valSet,
 	}), DefPartialResult4CountDistinctDurationSize + memDelta
@@ -212,7 +211,7 @@ func (e *countOriginalWithDistinct4Duration) AllocPartialResult() (pr PartialRes
 
 func (e *countOriginalWithDistinct4Duration) ResetPartialResult(pr PartialResult) {
 	p := (*partialResult4CountDistinctDuration)(pr)
-	p.valSet, _ = set.NewInt64SetWithMemoryUsage()
+	p.valSet, _ = newInt64SetWithMemoryUsage()
 }
 
 func (e *countOriginalWithDistinct4Duration) AppendFinalResult2Chunk(sctx sessionctx.Context, pr PartialResult, chk *chunk.Chunk) error {
@@ -244,7 +243,7 @@ func (e *countOriginalWithDistinct4Duration) UpdatePartialResult(sctx sessionctx
 }
 
 type partialResult4CountDistinctString struct {
-	valSet set.StringSetWithMemoryUsage
+	valSet stringSetWithMemoryUsage
 }
 
 type countOriginalWithDistinct4String struct {
@@ -252,7 +251,7 @@ type countOriginalWithDistinct4String struct {
 }
 
 func (e *countOriginalWithDistinct4String) AllocPartialResult() (pr PartialResult, memDelta int64) {
-	valSet, memDelta := set.NewStringSetWithMemoryUsage()
+	valSet, memDelta := newStringSetWithMemoryUsage()
 	return PartialResult(&partialResult4CountDistinctString{
 		valSet: valSet,
 	}), DefPartialResult4CountDistinctStringSize + memDelta
@@ -260,7 +259,7 @@ func (e *countOriginalWithDistinct4String) AllocPartialResult() (pr PartialResul
 
 func (e *countOriginalWithDistinct4String) ResetPartialResult(pr PartialResult) {
 	p := (*partialResult4CountDistinctString)(pr)
-	p.valSet, _ = set.NewStringSetWithMemoryUsage()
+	p.valSet, _ = newStringSetWithMemoryUsage()
 }
 
 func (e *countOriginalWithDistinct4String) AppendFinalResult2Chunk(sctx sessionctx.Context, pr PartialResult, chk *chunk.Chunk) error {
@@ -299,11 +298,11 @@ type countOriginalWithDistinct struct {
 }
 
 type partialResult4CountWithDistinct struct {
-	valSet set.StringSetWithMemoryUsage
+	valSet stringSetWithMemoryUsage
 }
 
 func (e *countOriginalWithDistinct) AllocPartialResult() (pr PartialResult, memDelta int64) {
-	valSet, memDelta := set.NewStringSetWithMemoryUsage()
+	valSet, memDelta := newStringSetWithMemoryUsage()
 	return PartialResult(&partialResult4CountWithDistinct{
 		valSet: valSet,
 	}), DefPartialResult4CountWithDistinctSize + memDelta
@@ -311,7 +310,7 @@ func (e *countOriginalWithDistinct) AllocPartialResult() (pr PartialResult, memD
 
 func (e *countOriginalWithDistinct) ResetPartialResult(pr PartialResult) {
 	p := (*partialResult4CountWithDistinct)(pr)
-	p.valSet, _ = set.NewStringSetWithMemoryUsage()
+	p.valSet, _ = newStringSetWithMemoryUsage()
 }
 
 func (e *countOriginalWithDistinct) AppendFinalResult2Chunk(sctx sessionctx.Context, pr PartialResult, chk *chunk.Chunk) error {
