+++ executor/aggfuncs/func_varpop.go
@@ -19,7 +19,6 @@ import (
 	"github.com/pingcap/errors"
 	"github.com/pingcap/tidb/sessionctx"
 	"github.com/pingcap/tidb/util/chunk"
-	"github.com/pingcap/tidb/util/set"
 )
 
 const (
@@ -126,7 +125,7 @@ type partialResult4VarPopDistinctFloat64 struct {
 	count    int64
 	sum      float64
 	variance float64
-	valSet   set.Float64SetWithMemoryUsage
+	valSet   float64SetWithMemoryUsage
 }
 
 func (e *varPop4DistinctFloat64) AllocPartialResult() (pr PartialResult, memDelta int64) {
@@ -134,7 +133,7 @@ func (e *varPop4DistinctFloat64) AllocPartialResult() (pr PartialResult, memDelt
 	p.count = 0
 	p.sum = 0
 	p.variance = 0
-	p.valSet, memDelta = set.NewFloat64SetWithMemoryUsage()
+	p.valSet, memDelta = newFloat64SetWithMemoryUsage()
 	return PartialResult(p), DefPartialResult4VarPopDistinctFloat64Size + memDelta
 }
 
@@ -143,7 +142,7 @@ func (e *varPop4DistinctFloat64) ResetPartialResult(pr PartialResult) {
 	p.count = 0
 	p.sum = 0
 	p.variance = 0
-	p.valSet, _ = set.NewFloat64SetWithMemoryUsage()
+	p.valSet, _ = newFloat64SetWithMemoryUsage()
 }
 
 func (e *varPop4DistinctFloat64) AppendFinalResult2Chunk(sctx sessionctx.Context, pr PartialResult, chk *chunk.Chunk) error {
