+++ util/set/string_set.go
@@ -51,44 +51,3 @@ func (s StringSet) Intersection(rhs StringSet) StringSet {
 func (s StringSet) Count() int {
 	return len(s)
 }
-
-const (
-	// DefStringSetBucketMemoryUsage = bucketSize*(1+unsafe.Sizeof(string) + unsafe.Sizeof(struct{}))+2*ptrSize
-	// ref https://github.com/golang/go/blob/go1.15.6/src/reflect/type.go#L2162.
-	// The bucket size may be changed by golang implement in the future.
-	DefStringSetBucketMemoryUsage = 8*(1+16+0) + 16
-	// Maximum average load of a bucket that triggers growth is 6.5.
-	// Represent as loadFactorNum/loadFactorDen, to allow integer math.
-	loadFactorNum = 13
-	loadFactorDen = 2
-)
-
-// StringSetWithMemoryUsage is a string set with memory usage.
-type StringSetWithMemoryUsage struct {
-	StringSet
-	bInMap int64
-}
-
-// NewStringSetWithMemoryUsage builds a string set.
-func NewStringSetWithMemoryUsage(ss ...string) (setWithMemoryUsage StringSetWithMemoryUsage, memDelta int64) {
-	set := make(StringSet, len(ss))
-	setWithMemoryUsage = StringSetWithMemoryUsage{
-		StringSet: set,
-		bInMap:    0,
-	}
-	memDelta = DefStringSetBucketMemoryUsage * (1 << setWithMemoryUsage.bInMap)
-	for _, s := range ss {
-		memDelta += setWithMemoryUsage.Insert(s)
-	}
-	return setWithMemoryUsage, memDelta
-}
-
-// Insert inserts `val` into `s` and return memDelta.
-func (s *StringSetWithMemoryUsage) Insert(val string) (memDelta int64) {
-	s.StringSet.Insert(val)
-	if s.Count() < (1<<s.bInMap)*loadFactorNum/loadFactorDen {
-		memDelta = DefStringSetBucketMemoryUsage * (1 << s.bInMap)
-		s.bInMap++
-	}
-	return memDelta + int64(len(val))
-}
